move POSReceiptEligibilityBanner to separate file
